refactor(tokens): move the stories that use vars into the vars package
@launchpad-ui/vars depends on @launchpad-ui/tokens, and three story files
in tokens imported { vars } from @launchpad-ui/vars. That is a cycle in
the import graph: tokens -> vars -> tokens. It does not fail an install
today only because tokens never declares vars in its manifest, so the
cycle is undeclared rather than absent. Any install strategy that
declares dependencies honestly and rejects cycles reports it.

color.stories.tsx, spacing.stories.tsx and typography.stories.tsx move to
packages/vars/stories/, where the dependency on tokens already exists and
points the right way. Their relative imports of the tokens package are
repointed (../dist, ../tokens/spacing.json, ./colorTokens). The story
titles Tokens/Color, Tokens/Spacing and Tokens/Typography are unchanged,
so the Storybook navigation and the snapshot identities are unchanged.

colorTokens.tsx stays in packages/tokens/stories next to its test, and
the moved color story imports it by relative path. Imports of
../../components/src resolve to the same files from the new directory,
since both directories sit at the same depth.
