+++ packages/vars/stories/color.stories.tsx
@@ -5,8 +5,7 @@ import { vars } from '@launchpad-ui/vars';
 import { Button } from '../../components/src/Button';
 import { ToastRegion, toastQueue } from '../../components/src/Toast';
 import { Tooltip, TooltipTrigger } from '../../components/src/Tooltip';
-
-import { type ComputedValue, getTokenHex, getTokenValue, TokenCode } from './colorTokens';
+import { type ComputedValue, getTokenHex, getTokenValue, TokenCode } from '../../tokens/stories/colorTokens';
 
 export default {
 	title: 'Tokens/Color',
